# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripio",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/persons": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Вывод списка записей (index())",
                "description": "Доступ только авторизованным пользователям",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    },
                    "200": {
                        "description": "Everything OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Persons"
                ],
                "summary": "Добавление новой записи (store())",
                "description": "Возвращает новый созданный город",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/persons/{id}": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Вывод конкретной записи (show())",
                "description": "Returns person data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Persons"
                ],
                "summary": "Обновление записи (update())",
                "description": "Returns updated person data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Persons"
                ],
                "summary": "Удаление записи (delete())",
                "description": "Deletes a record and returns no content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PersonSchema": {
                "properties": {
                    "id": {
                        "description": "Person ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "hotel_id": {
                        "description": "Foreign key hotel",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "name": {
                        "description": "Customer Name",
                        "type": "string",
                        "example": "Иван",
                        "nullable": false
                    },
                    "lastname": {
                        "description": "Customer lasname",
                        "type": "string",
                        "example": "Иванов",
                        "nullable": false
                    },
                    "patronymic": {
                        "description": "Customer patronymic",
                        "type": "string",
                        "example": "Иванович",
                        "nullable": false
                    },
                    "passport": {
                        "description": "Customer passport",
                        "type": "string",
                        "example": "4524761611",
                        "nullable": false
                    },
                    "email": {
                        "description": "Customer email",
                        "type": "string",
                        "example": "example@test.ru",
                        "nullable": false
                    },
                    "phone": {
                        "description": "Customer phone",
                        "type": "string",
                        "example": "79999999999",
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Date of person creation",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Date of last updating person data",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}